requirements.txt
# Minimal Python dependencies for training
numpy
Pillow
torch
torchvision
tqdm
